refactor: useHooks 파일 통합
