+++ src/components/bookNote/periNote/PeriNote.tsx
@@ -3,8 +3,8 @@ import { useOutletContext } from "react-router-dom";
 import styled from "styled-components";
 
 import { PeriNoteData, PreNoteData } from "../../../pages/BookNote";
-import { patchBookNote, useFetchNote } from "../../../utils/lib/bookNote";
 import { deepCopyTree, getNodeByPath } from "../../../utils/tree";
+import { patchBookNote, useFetchNote } from "../../../utils/useHooks";
 import { Error404, Loading } from "../../common";
 import { Button } from "../../common/styled/Button";
 import { Complete, ExButton, PeriModal, PriorQuestion, StepUp } from "..";
