+++ src/components/bookNote/preNote/PreNote.tsx
@@ -4,7 +4,7 @@ import styled, { css } from "styled-components";
 
 import { PeriNoteData, PreNoteData } from "../../../pages/BookNote";
 import { PeriNoteTreeNode } from "../../../utils/dataType";
-import { patchBookNote, useFetchNote } from "../../../utils/lib/bookNote";
+import { patchBookNote, useFetchNote } from "../../../utils/useHooks";
 import { Error404, Loading } from "../../common";
 import { Button } from "../../common/styled/Button";
 import { PopUpPreDone, PreNoteForm, QuestionThree } from "..";
