+++ src/utils/dataType.ts
@@ -35,3 +35,16 @@ interface AnswerThree {
   answerThree: PeriNoteTreeNode;
   reviewSt: number;
 }
+
+export interface PeriNoteData {
+  answerThree: PeriNoteTreeNode;
+  reviewSt: number;
+}
+
+export interface PreNoteData {
+  answerOne: string;
+  answerTwo: string;
+  questionList: string[];
+  reviewSt: number;
+  finishSt?: boolean;
+}
