+++ src/utils/useHooks.ts
@@ -1,8 +1,11 @@
+import axios from "axios";
 import { useEffect, useState } from "react";
 import { useRecoilState } from "recoil";
 
 import { BookcaseInfo } from "../pages/Bookcase";
 import { isLoginState } from "./atoms";
+import { PeriNoteData, PreNoteData } from "./dataType";
+import { client } from "./lib";
 import { getData } from "./lib/api";
 
 export const useGetBookcase = (key: string) => {
@@ -85,3 +88,48 @@ export function useCheckLoginState() {
 
   return { isLogin, isLoginLoading };
 }
+
+export function useFetchNote<T>(token: string, key: string, initialState: T) {
+  const [data, setData] = useState<T>(initialState);
+  const [isLoading, setIsLoading] = useState<boolean>(true);
+  const [isError, setIsError] = useState<boolean>(false);
+
+  useEffect(() => {
+    (async function () {
+      try {
+        const {
+          data: { data },
+        } = await client(token).get(key);
+
+        if (data.questionList && !data.questionList.length) {
+          // 버그? 사용자가 질문 리스트를 모두 지우고 저장해도 다시 빈 input이 생성됨
+          // 처음 추가된 책의 review에 대해서 서버에서 questionList를 []가 아닌 [""]로 주면 해결될 듯
+          setData({ ...data, questionList: [""] });
+        } else {
+          setData(data);
+        }
+      } catch (err) {
+        if (axios.isAxiosError(err)) {
+          // isError 처리에 대하여 고민해보기
+          setIsError(true);
+        }
+      } finally {
+        setIsLoading(false);
+      }
+    })();
+  }, []);
+
+  return { data, setData, isLoading, isError };
+}
+
+export const patchBookNote = async (token: string, key: string, body: PreNoteData | PeriNoteData) => {
+  try {
+    const { data } = await client(token).patch(key, body);
+
+    return data.data;
+  } catch (err) {
+    if (axios.isAxiosError(err)) {
+      return err.response;
+    }
+  }
+};
